requirements.txt
urwid==2.1.0
aiohttp==3.6.2
python-socketio==4.4.0
pylint